+++ CHANGES.rst
@@ -6,6 +6,9 @@ dev (master)
 
 * Change ``is_ipaddress`` to not detect IPvFuture addresses. (Pull #1583)
 
+* Change ``parse_url`` to percent-encode invalid characters within the
+  path, query, and target components. (Pull #1586)
+
 
 1.25.1 (2019-04-24)
 -------------------